Update install_guacamole.sh
Changed the regex to adapt to changes in the default nginx site config and added a wait to make sure the file is actually created by the container before trying to make changes.  Also removed the changes to waiting for mysql as that mechanism has changed in the underlying script and is no longer necessary as it directly checks the health of the mysql container instead of waiting a set time.